v5.16.0 Kanua Kenten: Introduce SQL caching with File fallback
Provides a path for fwew-api and other fwew-lib clients to interact with
the fwedit db (using environment variables for the connection creds in
the client). This update is mostly for server side shenanigans. Regular
users will not be affected or notice any real difference due to falling
back on dictionary-v2.txt as before, by default.